feat: refactor SelectAssigneeModal to use MembersList component and add styles
