+++ src/presentation/widgets/tasks/MembersList/MembersList.module.css
@@ -1,0 +1,25 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  padding-top: 20px;
+  height: 100%;
+}
+
+.membersList {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  overflow-y: auto;
+  padding-bottom: 20px;
+}
+
+.userItem {
+  transition: background-color 0.3s ease;
+}
+
+.userItem:hover {
+  cursor: pointer;
+  background-color: var(--grey);
+  border-radius: 16px;
+}
